Created `db.go`
- Refactor the database logic out of `app.go`
